+++ testutil/fixtures/load/frozen.go
@@ -85,7 +85,7 @@ func FrozenImagesLinux(ctx context.Context, client client.APIClient, images ...s
 func imageExists(ctx context.Context, client client.APIClient, name string) bool {
 	ctx, span := otel.Tracer("").Start(ctx, "check image exists: "+name)
 	defer span.End()
-	_, _, err := client.ImageInspectWithRaw(ctx, name)
+	_, err := client.ImageInspect(ctx, name)
 	if err != nil {
 		span.RecordError(err)
 	}